Switch from SafeMath contract to SafeMath library

+++ contracts/assets/AragonToken.sol
@@ -12,7 +12,7 @@ contract AragonToken is PreminedAsset {
     // Constant token specific fields
     string public constant name = "Aragon Network Token";
     string public constant symbol = "ANT";
-    uint public constant decimals = 18;
+    uint8 public constant decimals = 18;
     uint public constant preminedAmount = 10*28;
 
     // NON-CONSTANT METHODS
