+++ contracts/assets/AventusToken.sol
@@ -12,7 +12,7 @@ contract AventusToken is PreminedAsset {
     // Constant token specific fields
     string public constant name = "AventCoin";
     string public constant symbol = "AVT";
-    uint public constant decimals = 18;
+    uint8 public constant decimals = 18;
     uint public constant preminedAmount = 10**28;
 
     // NON-CONSTANT METHODS
