+++ contracts/assets/EuroToken.sol
@@ -12,7 +12,7 @@ contract EuroToken is PreminedAsset {
     // Constant token specific fields
     string public constant name = "Euro Token";
     string public constant symbol = "EUR-T";
-    uint public constant decimals = 8;
+    uint8 public constant decimals = 8;
     uint public constant preminedAmount = 10**28;
 
     // NON-CONSTANT METHODS
