+++ contracts/assets/MelonToken.sol
@@ -12,7 +12,7 @@ contract MelonToken is PreminedAsset {
     // Constant token specific fields
     string public constant name = "Melon Token";
     string public constant symbol = "MLN-T";
-    uint public constant decimals = 18;
+    uint8 public constant decimals = 18;
     uint public constant preminedAmount = 10**28;
 
     // NON-CONSTANT METHODS
@@ -20,5 +20,5 @@ contract MelonToken is PreminedAsset {
     function MelonToken()
         PreminedAsset(name, symbol, decimals, preminedAmount)
     {}
-      
+
 }
